Kadr 3

+++ app/Http/Controllers/AdminController.php
@@ -3,34 +3,46 @@
 namespace App\Http\Controllers;
 
 use App\Article;
+use Illuminate\Contracts\Foundation\Application;
 use Illuminate\Contracts\View\Factory;
+use Illuminate\Http\RedirectResponse;
 use Illuminate\Http\Request;
 use Illuminate\Http\Response;
 use Illuminate\Support\Facades\DB;
 use Illuminate\Support\Facades\Auth;
+use Illuminate\Validation\UnauthorizedException;
 use Illuminate\View\View;
 use App\User;
 
 class AdminController extends Controller
 {
     /**
-     * @param Request $request
+     * @const array
+     * IDs of roles_users entities eligible for accessing admin panel
+     */
+    public const ADMIN_ACCESS = [1];
+
+    /**
      * @return Factory|View
      */
-    public function dashboard(Request $request){
-        $news = Article::whereRaw(1)->orderBy('id', 'DESC')->paginate(10);
-        return view('admin.dashboard', compact('news'));
+    public function dashboard(){
+        $articles = Article::query()->orderBy('id', 'DESC')->paginate(10);
+        return view('admin.dashboard', compact('articles'));
     }
 
     /**
      * Store a newly created resource in storage.
      *
      * @param Request $request
-     * @return Response
+     * @return RedirectResponse
      */
 
     public function storeArticle(Request $request)
     {
+        if (!Auth::check()) {
+            throw new UnauthorizedException();
+        }
+
         $request->validate([
             'title' => 'required',
             'text' => 'required',
@@ -57,11 +69,15 @@ class AdminController extends Controller
     /**
      * Show the form for creating a new resource.
      *
-     * @return Response
+     * @return Application|Factory|Response|View
      */
 
     public function createArticle()
     {
+        if (!Auth::check()) {
+            throw new UnauthorizedException();
+        }
+
         return view('article.create');
     }
 
@@ -69,32 +85,41 @@ class AdminController extends Controller
      * Show the form for editing the specified resource.
      *
      * @param int $id
-     * @return Response
+     * @return Application|Factory|Response|View
      */
 
     public function editArticle(int $id)
     {
-        $article = Article::find($id);
+        if (!Auth::check()) {
+            throw new UnauthorizedException();
+        }
+
+        $article = Article::findOrFail($id);
 
         return view('article.edit',compact('article'));
     }
 
 
-
     /**
      * Update the specified resource in storage.
      *
-     * @param  Request $request
-     * @param Article $article
-     * @return Response
+     * @param Request $request
+     * @param int $id
+     * @return RedirectResponse
      */
 
-    public function updateArticle(Request $request, Article $article)
+    public function updateArticle(Request $request, int $id)
     {
+        if (!Auth::check()) {
+            throw new UnauthorizedException();
+        }
+
         $request->validate([
             'name' => 'required',
             'detail' => 'required',
         ]);
+
+        $article = Article::findOrFail($id);
 
         $article->update($request->all());
 
@@ -105,14 +130,18 @@ class AdminController extends Controller
     /**
      * Remove the specified resource from storage.
      *
-     * @param Article $article
-     * @return Response
+     * @param int $id
+     * @return RedirectResponse
      *
-     * @throws \Throwable
      */
 
-    public function destroyArticle(Article $article)
+    public function destroyArticle(int $id)
     {
+        if (!Auth::check()) {
+            throw new UnauthorizedException();
+        }
+
+        $article = Article::findOrFail($id);
         $article->delete();
 
         return redirect()->route('admin.dashboard')
@@ -120,21 +149,24 @@ class AdminController extends Controller
     }
 
     /**
-     * @param Request $request
      * @return Factory|View
      */
-    public function usersDashboard(Request $request){
+    public function usersDashboard(){
+        if (!Auth::check()) {
+            throw new UnauthorizedException();
+        }
+
         $users = User::whereRaw(1)->orderBy('id', 'DESC')->paginate(10);
         return view('admin.usersDashboard', compact('users'));
     }
 
     /**
-     * @param User $user
+     * @param int $id
      * @return Factory|View
      */
-    public function editUser(User $user){
-        $id = $user->id;
+    public function editUser(int $id){
+        $user = User::findOrFail($id);
 
-        return view('editUser', compact('id'));
+        return view('editUser', compact('user'));
     }
 }
